Incorporating in Sherwood relation calculations
aLMOST GOT RID OF APRTITION COEFFECEINT FROM Dynamics, but then remembered that it was included in non-dimensionalizing unbound nanoparticle concentration so the boundary condition is 1

Got to point of running code, but no cigar on residual going to zero. first pass at residual calculation was very close to zero so I think it is a JAcobian problem now. Many changes to underlying math since many errors found.