end of day Nov 3
Almost have model vs literature results lined up. Weird problems with time=14 and 28 minutes, but otherwise I think the script is almost ready. Two things to add to script: ensuring x=1 is biofilm surface (currently it is about x=0.9 because of nature of literature data), and need to plot mode vs literature values. The other thing that needs to be done (not in this script) is how to make my boundary condition a function of time. If I just make the boundary condition a function of time in the code, will that have other downstream effects? Will need to just sit down and ponder that one before making the change.